fix(cloudscape-website): implement lazilyRender which wraps any unresolved tokens in CfnJson
When providing a runtime config with dynamic values, this gets synthesized into an asset with
markers. This is fine for the first deployment, however once a user changes a value the marker
doesn't change and Cloudformation doesn't think there are any changes. This fix ensures that any
changes to the runtimeConfig result in a deployment.

fix #22