feat(license): sync package.json, pyproject.toml and license files in sync generator
Instead of writing the root LICENSE file once in the license generator (and synchronising
subprojects to match the root license), we synchronise the root LICENSE and all subproject LICENSE
files to the license configured in nx plugin config. Additionally, move the package.json update to
the sync generator, and add support for updating pyproject.toml files.

Fixes #42