fix(open-api): handle object schemas with additional properties and pattern properties
Schemas which mixed objects and additional properties were causing clients to be generated which did
not compile.

We'd render invalid code like `$IO.unknown.toJson(xxxx)`.

Additionally, `patternProperties` are part of the JSON schema version used by OpenAPI v3.1, and so
therefore should be handled too.

This change adds support for additional properties and pattern properties.

Note that due to a limitation of the TypeScript type system, we must type these as `unknown`, rather
than the type defined in the schema, for example this type is invalid:

```ts
export type SomeType = {
  name: string;
  [key: string]: number;
};
```

It's also not possible to define the `key` type as "any string except name" in TypeScript :(

While the generated types for additional properties and pattern properties have this limitation, we
still ensure that these properties are marshalled correctly.