build: run unit tests in parallel and break up ts-client tests
It seems as though the Nx issue that was causing our tests to fail when run in parallel has been
resolved, so long as the daemon is used, so we remove the single-threaded restrictions from config.
Additionally, split up the large ts-client test suite into logical groups which can be run in
parallel, and make them easier to digest.