feat(fastapi): generate an echo api by default for consistency with trpc
For tRPC we generate an echo procedure as an example to follow. For FastAPI we were generating a
read_root operation which can be confusing for users. We update to generate an echo route for
consistency with tRPC. Additionally, we define a model for the response returned by the unhandled
exception middleware for better error handling in the generated code when using the api-connection.

Fixes #92