chore: add usage metric identifiers to cloudformation stack descriptions
In order to help us improve the Nx Plugin for AWS, we collect metrics on generator usage. This
process involves no additional network requests, we simply leverage existing AWS usage metrics. We
append an ID, version, and list of generator IDs to CloudFormation Stack descriptions. This is done
in the `MetricsAspect` in `common/constructs`.