feat: add ts#nx-generator generator for generating generators
Add a new generator which adds a generator to an existing typescript project.

When run within this repo, it adds some extra pieces specific to our generators
such as adding a guide to the docs.

Note that we remove the dist/xxx entry in the tsconfig.base.json paths since
this causes issues when running local generators, and doesn't seem to be
required for the build and/or IDE completion.

Fixes #127