feat: use type-safe per-operation lambda handlers for trpc and fastapi
This change updates both trpc and fastapi apis to use individual lambda functions for each operation
in the API, in order for customers to follow the principle of least privilege, granting the minimal
permissions for each lambda handler to perform its specific function. We achieve this in a type-safe
manner, which uses type inference for trpc and minimal openapi code generation for fastapi. We
provide options for deploying using either an API Gateway REST API or HTTP API.

Fixes #146