feat(apis): support for cognito auth in ts#trpc-api and py#fast-api
This change adds support for cognito auth for tRPC and FastAPI, for both REST and HTTP API
infrastructure. We remove the "auth" option from the api-connection generator and instead specify
the auth method in the API generators, since this determines the infrastructure set up. We update
the api-connection generator to infer the auth method from the target project. Additionally we
simplify the tRPC client provider to create a specific provider for each tRPC API, instead of
creating a shared provider creation function. This allows us to have different auth methods for
different tRPC APIs, whereas previously all would need to have the same auth method.

Fixes #161