fix(license): support workspace not in root git repo
The branch to honour .gitignores or not assumed that the workspace was
in the root of the git repo, ie a .git folder existed. Fall back to
asking git if we are within a repo to handle this case.

Additionally fix a bug where only root files were found in the non-git
case.

Fixes #199