fix(license): fix bug parsing lines that don't end a block comment but use block comment end syntax
We were incorrectly assuming that if a line ends with a block comment it ends the block comment!
This isn't the case, we need to be within a block comment to end it. Check that we're either within
a block comment or on a line that both starts and ends the block comment.

Fixes #200