feat(ts#lambda-function): upgrade to zod v4 in line with latest powertools version (#252)
* feat(ts#lambda-function): upgrade to zod v4 in line with latest powertools version

Zod v4 has been released and published to npm as 4.x (rather than as part of 3.x). This change
updates ts#lambda-function to use Zod v4 since lambda powertools now supports it.

This also fixes an issue where if the latest powertools was used with Zod 3, lambda functions
would fail due to the mismatching versions.

As well as this, we remove the `/v4` imports from trpc api and upgrade the dependency to Zod 4.

Note that `@modelcontextprotocol/sdk`` still requires zod v3, so we explicitly install it
for MCP only.

Fixes #251

* docs: update translations

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>