feat(ts#infra): wrap stack in a stage and support cross-stack runtime config (#280)
* feat(ts#infra): wrap stack in a stage and support cross-stack runtime config

Vend an ApplicationStage which consists of the ApplicationStack, making it easier to build
multi-stack applications. Associate the RuntimeConfig construct with the parent Stage if available,
otherwise fall back to the parent Stack. Additionally adjust the load:runtime-config target for the
website to be more robust by looking for stacks prefixed by the scope and outputs for the specific
website the target pertains to.

Fixes #162
Fixes #214

* docs: update translations

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>