fix(ts#smithy-api): fix first time serve-local bug (affected fastapi too)
Fix the bug where serve-local failed due to a missing client if you hadn't run an initial build.
Additionally add extensions.smithy as part of the api-connection rather than relying on users
reading the documentation to understand how to customise the client. Remove the index.ts with a
hello function from the generated smithy backend.