fix(py): multiple platform bundle targets for single project (#315)
* fix(py): multiple platform bundle targets for single project

When a py#lambda-function was added to the same project as a py#mcp-server (or vice-versa), the
existing bundle target was reused, but this caused a bundle to be deployed for the wrong platform.
To address this, we create a platform specific bundle target and ensure we always deploy the correct
bundle.

Fixes #308
---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>