fix(ts#mcp-server): upgrade zod and mcp sdk to compatible versions (#387)
Yarn and Bun use the hoisted version of zod which we render in the root package.json instead of the
version which the mcp sdk depends on. This was causing an issue where the mcp sdk required zod 4.2
and above but yarn was resolving the version we vended which was 4.1.